fix driver test, change eps's type double to float

+++ driver/layernorm_driver.hpp
@@ -147,7 +147,7 @@ private:
     std::vector<Tref> dweighthost;
     std::vector<Tref> dbiashost;
 
-    double eps;
+    float eps;
     int dim;
     miopenLayerNormMode_t mode;
 };
